Fixed Lion scrollbar issues, big API change.
This change implements scrollbar fixes from @caffeinatedapp. These fixes required the API to significantly change, for the better in my opinion. scrollviews and web views are now subclassed with LNScrollView and LNWebView.

This change also improves the API for setting the background pattern, which can now be set on each individual view.

fix #1